massive changes - added the ... ellipsis and selector
