+++ frontend/src/components/Worksheet/CalendarEvent.tsx
@@ -2,9 +2,7 @@ import React, { useCallback } from 'react';
 import chroma from 'chroma-js';
 import LinesEllipsis from 'react-lines-ellipsis';
 import responsiveHOC from 'react-lines-ellipsis/lib/responsiveHOC';
-import ColorPickerButton from './ColorPickerButton';
 import WorksheetHideButton from './WorksheetHideButton';
-import { WorksheetMoveDropdown } from './WorksheetMoveDropdown';
 import { useStore } from '../../store';
 import type { RBCEvent } from '../../utilities/calendar';
 import styles from './CalendarEvent.module.css';
