+++ frontend/src/components/Worksheet/WorksheetItemActionsButton.tsx
@@ -1,11 +1,29 @@
 import React, { useState, useRef } from 'react';
-import { Overlay, Popover, Tooltip, OverlayTrigger } from 'react-bootstrap';
+import {
+  Overlay,
+  Popover,
+  Tooltip,
+  OverlayTrigger,
+  Modal,
+  Button,
+  DropdownButton,
+  Dropdown,
+} from 'react-bootstrap';
 import { useShallow } from 'zustand/react/shallow';
 import { useStore } from '../../store';
 import type { RBCEvent } from '../../utilities/calendar';
 import { FaEllipsisH } from 'react-icons/fa';
-import ColorPickerButton from './ColorPickerButton';
-import { WorksheetMoveDropdown } from './WorksheetMoveDropdown';
+import { MdEdit, MdMoveToInbox } from 'react-icons/md';
+import { updateWorksheetCourses } from '../../queries/api';
+import { useWorksheetNumberOptions } from '../../slices/WorksheetSlice';
+import { Calendar } from 'react-big-calendar';
+import { HexColorPicker } from 'react-colorful';
+import chroma from 'chroma-js';
+import { CalendarEventBody, useEventStyle } from './CalendarEvent';
+import { localizer } from '../../utilities/calendar';
+import { worksheetColors } from '../../utilities/constants';
+import { SurfaceComponent, Input } from '../Typography';
+import styles from './ColorPickerButton.module.css';
 
 function WorksheetItemActionsButton({
   event,
@@ -21,22 +39,23 @@ function WorksheetItemActionsButton({
       viewedWorksheetNumber: state.viewedWorksheetNumber,
     })),
   );
-  const [open, setOpen] = useState(false);
+  const [popoverOpen, setPopoverOpen] = useState(false);
+  const [openColorPicker, setOpenColorPicker] = useState(false);
+  const [openWorksheetMove, setOpenWorksheetMove] = useState(false);
   const targetRef = useRef<HTMLButtonElement>(null);
 
-  const toggleDropdown = (e: React.MouseEvent) => {
+  const togglePopover = (e: React.MouseEvent) => {
     e.stopPropagation();
     e.preventDefault();
-    setOpen((prev) => !prev);
+    setPopoverOpen((prev) => !prev);
   };
 
-  const closeDropdown = () => {
-    setOpen(false);
+  const closePopover = () => {
+    setPopoverOpen(false);
   };
 
   return (
     <div
-      // Prevent click events from bubbling up to parent components
       onClick={(e) => {
         e.stopPropagation();
         e.preventDefault();
@@ -55,7 +74,7 @@ function WorksheetItemActionsButton({
           type="button"
           className={className}
           ref={targetRef}
-          onClick={toggleDropdown}
+          onClick={togglePopover}
           aria-label="Item actions"
         >
           <FaEllipsisH color="var(--color-text-dark)" />
@@ -64,11 +83,11 @@ function WorksheetItemActionsButton({
 
       <Overlay
         target={targetRef.current}
-        show={open}
+        show={popoverOpen}
         placement="bottom"
         containerPadding={20}
         rootClose
-        onHide={closeDropdown}
+        onHide={closePopover}
       >
         {(props) => (
           <Popover id="popover-contained" {...props}>
@@ -80,14 +99,297 @@ function WorksheetItemActionsButton({
                   gap: '8px',
                 }}
               >
-                <ColorPickerButton event={event} className={className} />
-                <WorksheetMoveDropdown event={event} className={className} />
+                <OverlayTrigger
+                  placement="bottom"
+                  overlay={
+                    <Tooltip id="color-tooltip">
+                      <small>Change color</small>
+                    </Tooltip>
+                  }
+                >
+                  <button
+                    type="button"
+                    className={className}
+                    onClick={(e) => {
+                      e.stopPropagation();
+                      e.preventDefault();
+                      setOpenColorPicker(true);
+                      setPopoverOpen(false);
+                    }}
+                    aria-label="Change color"
+                  >
+                    <MdEdit color="var(--color-text-dark)" />
+                  </button>
+                </OverlayTrigger>
+                <OverlayTrigger
+                  placement="bottom"
+                  overlay={
+                    <Tooltip id="move-tooltip">
+                      <small>Move to another worksheet</small>
+                    </Tooltip>
+                  }
+                >
+                  <button
+                    type="button"
+                    className={className}
+                    onClick={(e) => {
+                      e.stopPropagation();
+                      e.preventDefault();
+                      setOpenWorksheetMove(true);
+                      setPopoverOpen(false);
+                    }}
+                    aria-label="Move course"
+                  >
+                    <MdMoveToInbox color="var(--color-text-dark)" />
+                  </button>
+                </OverlayTrigger>
               </div>
             </Popover.Body>
           </Popover>
         )}
       </Overlay>
+
+      {openColorPicker && (
+        <ColorPickerModal
+          event={event}
+          className={className}
+          onClose={() => setOpenColorPicker(false)}
+        />
+      )}
+
+      {openWorksheetMove && (
+        <WorksheetMoveModal
+          event={event}
+          className={className}
+          onClose={() => setOpenWorksheetMove(false)}
+        />
+      )}
     </div>
+  );
+}
+
+function ColorPickerModal({
+  event,
+  className,
+  onClose,
+}: {
+  readonly event: RBCEvent;
+  readonly className?: string;
+  readonly onClose: () => void;
+}) {
+  const worksheetsRefresh = useStore((state) => state.worksheetsRefresh);
+  const { viewedSeason, viewedWorksheetNumber } = useStore(
+    useShallow((state) => ({
+      viewedSeason: state.viewedSeason,
+      viewedWorksheetNumber: state.viewedWorksheetNumber,
+    })),
+  );
+  const [newColor, setNewColor] = useState(event.color);
+
+  const handleClose = () => {
+    setNewColor(event.color);
+    onClose();
+  };
+
+  return (
+    <Modal show onHide={handleClose} centered>
+      <Modal.Body className={styles.modalBody}>
+        <Picker color={newColor} setColor={setNewColor} />
+        <Preview event={event} color={newColor} />
+      </Modal.Body>
+      <Modal.Footer>
+        <Button variant="secondary" onClick={handleClose}>
+          Close
+        </Button>
+        <Button
+          variant="primary"
+          onClick={async () => {
+            await updateWorksheetCourses({
+              action: 'update',
+              season: viewedSeason,
+              crn: event.listing.crn,
+              worksheetNumber: viewedWorksheetNumber,
+              color: newColor,
+            });
+            await worksheetsRefresh();
+            onClose();
+          }}
+        >
+          Save changes
+        </Button>
+      </Modal.Footer>
+    </Modal>
+  );
+}
+
+function WorksheetMoveModal({
+  event,
+  className,
+  onClose,
+}: {
+  readonly event: RBCEvent;
+  readonly className?: string;
+  readonly onClose: () => void;
+}) {
+  const { viewedSeason, viewedWorksheetNumber, worksheetsRefresh } = useStore(
+    (state) => ({
+      viewedSeason: state.viewedSeason,
+      viewedWorksheetNumber: state.viewedWorksheetNumber,
+      worksheetsRefresh: state.worksheetsRefresh,
+    }),
+  );
+  const options = useWorksheetNumberOptions('me', viewedSeason);
+  const filteredOptions = Object.values(options).filter(
+    (option) => option.value !== viewedWorksheetNumber,
+  );
+
+  const handleSelect = async (worksheetKey: string | null) => {
+    if (!worksheetKey) return;
+    const newWorksheetNumber = Number(worksheetKey);
+
+    await updateWorksheetCourses({
+      action: 'remove',
+      season: viewedSeason,
+      crn: event.listing.crn,
+      worksheetNumber: viewedWorksheetNumber,
+    });
+
+    const addResult = await updateWorksheetCourses({
+      action: 'add',
+      season: viewedSeason,
+      crn: event.listing.crn,
+      worksheetNumber: newWorksheetNumber,
+      color: event.color,
+      hidden: false,
+    });
+
+    if (addResult) {
+      worksheetsRefresh();
+      onClose();
+    }
+  };
+
+  return (
+    <Modal show onHide={onClose} centered>
+      <Modal.Body className={styles.modalBody}>
+        <DropdownButton
+          title="Select Worksheet"
+          onSelect={handleSelect}
+          variant="secondary"
+          size="sm"
+        >
+          {filteredOptions.map(({ value, label }) => (
+            <Dropdown.Item key={value} eventKey={String(value)}>
+              {label}
+            </Dropdown.Item>
+          ))}
+        </DropdownButton>
+      </Modal.Body>
+      <Modal.Footer>
+        <Button variant="secondary" onClick={onClose}>
+          Close
+        </Button>
+      </Modal.Footer>
+    </Modal>
+  );
+}
+
+function ColorInput({
+  color,
+  setColor,
+}: {
+  readonly color: string;
+  readonly setColor: (newColor: string) => void;
+}) {
+  const [invalid, setInvalid] = useState(false);
+  const [value, setValue] = useState(color);
+  const [prevColor, setPrevColor] = useState(color);
+  if (color !== prevColor) {
+    setValue(color);
+    setPrevColor(color);
+  }
+  return (
+    <Input
+      type="text"
+      value={value}
+      isInvalid={invalid}
+      onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+        const newColor = e.target.value;
+        if (chroma.valid(newColor)) {
+          setColor(newColor);
+          setInvalid(false);
+        } else {
+          setInvalid(true);
+        }
+        setValue(newColor);
+      }}
+    />
+  );
+}
+
+function Picker({
+  color,
+  setColor,
+}: {
+  readonly color: string;
+  readonly setColor: (newColor: string) => void;
+}) {
+  return (
+    <div className={styles.pickerPanel}>
+      <ColorInput color={color} setColor={setColor} />
+      <HexColorPicker color={color} onChange={setColor} />
+      <div className={styles.presetColors}>
+        {worksheetColors.map((presetColor) => (
+          <button
+            type="button"
+            key={presetColor}
+            className={styles.presetColor}
+            style={{ background: presetColor }}
+            onClick={(e) => {
+              e.stopPropagation();
+              e.preventDefault();
+              setColor(presetColor);
+            }}
+            aria-label={`Set color to ${presetColor}`}
+          />
+        ))}
+      </div>
+    </div>
+  );
+}
+
+function Preview({
+  event,
+  color,
+}: {
+  readonly event: RBCEvent;
+  readonly color: string;
+}) {
+  const eventStyleGetter = useEventStyle();
+  const tempEvent = { ...event, color };
+  const start = new Date(tempEvent.start);
+  if (start.getMinutes() === 0) start.setHours(start.getHours() - 1);
+  start.setMinutes(0);
+  const end = new Date(tempEvent.end);
+  end.setHours(end.getHours() + 1);
+  end.setMinutes(0);
+  return (
+    <SurfaceComponent className={styles.eventPreview}>
+      <Calendar
+        defaultView="day"
+        views={['day']}
+        events={[tempEvent]}
+        date={tempEvent.start}
+        min={start}
+        max={end}
+        localizer={localizer}
+        toolbar={false}
+        components={{ event: CalendarEventBody }}
+        eventPropGetter={eventStyleGetter}
+        tooltipAccessor={undefined}
+        onNavigate={() => {}}
+      />
+    </SurfaceComponent>
   );
 }
